Preparing for update to use new call frame iterators
